+++ M06-AD/UF2/Activities/Act1/PracticaJDBC/src/main/java/daniel/Cabrera/classes/Client.java
@@ -30,13 +30,13 @@ public class Client {
                     viewClients(connection);
                     break;
                 case 4:
-                    Tarjeta.altaTarjeta(connection);
+                    Card.altaCard(connection);
                     break;
                 case 5:
-                    Tarjeta.bajaTarjeta(connection);
+                    Card.bajaCard(connection);
                     break;
                 case 6:
-                    Tarjeta.viewTarjeta(connection);
+                    Card.viewCard(connection);
                     break;
                 case 0:
                     submenu = 1;
@@ -58,17 +58,36 @@ public class Client {
         String clientSurname2 = scanner.nextLine();
         System.out.print("Introduce la regió del client: ");
         String clientRegion = scanner.nextLine();
-        String sql = "INSERT INTO CLIENT (clientID, name, surname1, surname2, region) VALUES (?, ?, ?, ?, ?)";
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, clientID);
-            statement.setString(2, clientName);
-            statement.setString(3, clientSurname1);
-            statement.setString(4, clientSurname2);
-            statement.setString(5, clientRegion);
-            statement.executeUpdate();
-            System.out.println("Client dado de alta correctamente.");
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+
+        while (true) {
+            String sqlCheckArtist = "SELECT * FROM CLIENT WHERE clientID = ?";
+            try (PreparedStatement statementCheckClient = connection.prepareStatement(sqlCheckArtist)) {
+                statementCheckClient.setString(1, clientID);
+                ResultSet resultSet = statementCheckClient.executeQuery();
+                if (!resultSet.next()) {
+                    String sql = "INSERT INTO CLIENT (clientID, name, surname1, surname2, region) VALUES (?, ?, ?, ?, ?)";
+                    try (PreparedStatement statement = connection.prepareStatement(sql)) {
+                        statement.setString(1, clientID);
+                        statement.setString(2, clientName);
+                        statement.setString(3, clientSurname1);
+                        statement.setString(4, clientSurname2);
+                        statement.setString(5, clientRegion);
+                        statement.executeUpdate();
+                        System.out.println("Client dado de alta correctamente.");
+                    } catch (SQLException e) {
+                        throw new RuntimeException(e);
+                    }
+                } else {
+                    System.out.println("Ya existe un client amb aquest codi.");
+                    System.out.print("Introduce el código del client. (prem -1 per sortir):");
+                    clientID = scanner.nextLine();
+                    if (clientID.equals("-1")) {
+                        return;
+                    }
+                }
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
         }
     }
 
@@ -77,13 +96,32 @@ public class Client {
         System.out.println("######################### BAIXA CLIENT #########################");
         System.out.println("Introduce el código del client que quieres dar de baja: ");
         String clientID = scanner.nextLine();
-        String sql = "DELETE FROM CLIENT WHERE clientID = ?";
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, clientID);
-            statement.executeUpdate();
-            System.out.println("Client dado de baja correctamente.");
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+
+        while (true) {
+            String sqlCheckArtist = "SELECT * FROM CLIENT WHERE clientID = ?";
+            try (PreparedStatement statementCheckClient = connection.prepareStatement(sqlCheckArtist)) {
+                statementCheckClient.setString(1, clientID);
+                ResultSet resultSet = statementCheckClient.executeQuery();
+                if (!resultSet.next()) {
+                    System.out.println("No existeix ningún client amb aquest codi.");
+                    System.out.println("Introduce el código del client que quieres dar de baja. (prem -1 per sortir):");
+                    clientID = scanner.nextLine();
+                    if (clientID.equals("-1")) {
+                        return;
+                    }
+                } else {
+                    String sql = "DELETE FROM CLIENT WHERE clientID = ?";
+                    try (PreparedStatement statement = connection.prepareStatement(sql)) {
+                        statement.setString(1, clientID);
+                        statement.executeUpdate();
+                        System.out.println("Client dado de baja correctamente.");
+                    } catch (SQLException e) {
+                        throw new RuntimeException(e);
+                    }
+                }
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
         }
     }
 
@@ -93,7 +131,7 @@ public class Client {
             ResultSet resultSet = statement.executeQuery();
 
             // Verificar si no hay filas en el resultSet
-            if (!resultSet.isBeforeFirst()) {
+            if (!resultSet.next()) {
                 System.out.println("No existen clients.");
             } else {
                 int i = 1;
@@ -111,7 +149,7 @@ public class Client {
                     System.out.println("Client Surname1: " + clientSurname1);
                     System.out.println("Client Surname2: " + clientSurname2);
                     System.out.println("Client Region: " + clientRegion);
-                    System.out.println("Client Cards: " + Tarjeta.viewTarjeta(connection, clientID));
+                    System.out.println("Client Cards: " + Card.viewCard(connection, clientID));
                     System.out.println();
                     i++;
                 }
